Move test to accomodate function move
